requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# For parquet support
pyarrow>=14.0.0

# For progress bars
tqdm>=4.65.0

# For S3 flat files
boto3>=1.28.0

# For live dashboard
streamlit>=1.28.0
matplotlib>=3.7.0

